240416-house(UI,NormalMon)

+++ Assets/1.Scripts/ArrowSc/ArrowDamageScriptable.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class ArrowDamageScriptable : ScriptableObject
 {
     public int initialDamage;
-    public float damagePerSecond;
+    public int damagePerSecond;
     public float duration;
     public string effectType;
     public ParticleSystem particleEffect;
